move checks to page
